requirements.txt

fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.35
pydantic==2.9.2
pydantic[email]
pyjwt==2.9.0
bcrypt==4.2.0
slowapi==0.1.9
bleach==6.1.0
python-dotenv==1.0.1
python-multipart==0.0.12
httpx
pytest
 